+++ leetcode/unique_email_address/copy1.py
@@ -2,33 +2,20 @@ from typing import List
 
 class Solution:
     def numUniqueEmails(self, emails: List[str]) -> int:
-        def normalize(before):
-            normalized = []
-            i = 0
-            while i < len(before):
-                char = before[i]
-                if char == '@':
-                    while i < len(before):
-                        normalized.append(before[i])
-                        i += 1
+        def canonicalize_email(before):
+            after = []
+            local, domain = before.split("@")
+            for c in local:
+                if c == "+":
                     break
+                if c == ".":
+                    continue
+                after.append(c)
+            after.append("@")
+            after.append(domain)
+            return "".join(after)
 
-                if char == '.':
-                    i += 1
-                    continue
-                
-                if char == '+':
-                    while i < len(before) and before[i] != '@':
-                        i += 1
-                    continue
-                
-                normalized.append(char)
-                i += 1
-            return "".join(normalized)
-                
-        
-        normalized_emails = set()
+        unique_emails = set()
         for email in emails:
-            normalized_emails.add(normalize(email))
-
-        return len(normalized_emails)
+            unique_emails.add(canonicalize_email(email))
+        return len(unique_emails)
