+++ leetcode/unique_email_address/memo.md
@@ -1,0 +1,6 @@
+問題：https://leetcode.com/problems/unique-email-addresses/description/
+
+
+## Step1
+- 一目見て解けそうだったので解いてみた
+- フラグを使って1文字ずつ処理する方法と、特殊な文字が来たときに、その文字に対応するルールをまとめ適用するか判断が分かれそう
